examples/sreality-monitor/requirements.txt
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
matplotlib>=3.7.0
